refactor: useState 외 필요 없는 파일 삭제
