+++ src/App.tsx
@@ -1,5 +1,21 @@
-function App() {
-  return null;
-}
+import { ChakraProvider, extendTheme } from '@chakra-ui/react';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { Provider } from 'react-redux';
+import { RouterProvider } from 'react-router-dom';
 
-export default App;
+import router from '@utils/Router';
+
+const theme = extendTheme({
+  colors: {
+    brand: {
+      main: '#789BFB',
+    },
+  },
+});
+
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
+  <ChakraProvider theme={theme}>
+    <RouterProvider router={router} />
+  </ChakraProvider>
+);
