--- src/main.tsx
@@ -1,24 +1,0 @@
-import { ChakraProvider, extendTheme } from '@chakra-ui/react';
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import { Provider } from 'react-redux';
-import { RouterProvider } from 'react-router-dom';
-import store from 'store/store';
-
-import router from '@utils/Router';
-
-const theme = extendTheme({
-  colors: {
-    brand: {
-      main: '#789BFB',
-    },
-  },
-});
-
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <Provider store={store}>
-    <ChakraProvider theme={theme}>
-      <RouterProvider router={router} />
-    </ChakraProvider>
-  </Provider>
-);
