+++ src/pages/Main.tsx
@@ -1,6 +1,5 @@
 import { Flex } from '@chakra-ui/react';
 
-import TestTravelProduct from '@components//TestTravelProduct';
 import Header from '@components/Header';
 import TravelProduct from '@components/TravelProduct';
 
@@ -8,8 +7,7 @@ function Main() {
   return (
     <Flex flexDirection={'column'} margin={10}>
       <Header />
-      {/* <TravelProduct /> */}
-      <TestTravelProduct />
+      <TravelProduct />
     </Flex>
   );
 }
